MEGA-UPDATE!
Updated Character Debug - mostly complete.

HP Up, HP Down, and Attack Test buttons are non-functional.

There is a bug when lowering Toughness that causes the game to crash to desktop if Toughness is lowered below 0, or Toughness is lowered to a point where MaxHP is less than or equal to 0.

New updatestats() sub added to the CharacterDebug form in order to prevent the Luck value from being re-randomized every time a stat is changed.