UPDATE!
Ranged weapon generation is coming together.  This will be revamped into the rest of the procedural item generation.

After that, room generation (with placeholders) and monster generation will be functional.

Combat mechanics need to be revamped - now that I know they're working, ranged weapons, armor, and augments have to be taken into account.

More to come.