semi-big update???
fixing combat functions.  ranged and melee for the player are accounted for, but active mob stuff needs to be set before mob combat can be done, which requires the targeting system to be done.

damage rolls and armor are also fixed. so there's that.

soon: targeting system, UI, active items/inventory